+++ packages/graphql/server/resolvers/defaultMutationResolvers.ts
@@ -9,11 +9,9 @@ import { getModel, getModelConnector } from "./context";
 import { throwError } from "./errors";
 
 import { ContextWithUser } from "./typings";
-import { ModelMutationPermissionsOptions } from "@vulcanjs/model";
 import { VulcanDocument } from "@vulcanjs/schema";
 import { MutationResolverDefinitions } from "../typings";
 import { VulcanGraphqlModel } from "../../typings";
-import { isMemberOf } from "@vulcanjs/permissions";
 
 const defaultOptions = {
   create: true,
@@ -27,56 +25,6 @@ const getUpdateMutationName = (typeName) => `update${typeName}`;
 const getDeleteMutationName = (typeName) => `delete${typeName}`;
 const getUpsertMutationName = (typeName) => `upsert${typeName}`;
 
-type OperationName = "create" | "update" | "delete";
-const operationChecks: {
-  [operationName in OperationName]: keyof ModelMutationPermissionsOptions;
-} = {
-  create: "canCreate",
-  update: "canUpdate",
-  delete: "canDelete",
-};
-
-interface MutationCheckOptions {
-  user?: any;
-  document?: VulcanDocument;
-  model: VulcanGraphqlModel;
-  context: any;
-  operationName: OperationName;
-}
-/*
-
-Perform security check before calling mutators
-
-*/
-export const performMutationCheck = (options: MutationCheckOptions) => {
-  const { user, document, model, context, operationName } = options;
-  const { typeName } = model.graphql;
-  const { Users } = context;
-  const permissionsCheck = model.permissions?.[operationChecks[operationName]];
-  let allowOperation = false;
-  const fullOperationName = `${typeName}:${operationName}`;
-  const documentId = document?._id;
-  const data = { documentId, operationName: fullOperationName };
-  // 1. if no permission has been defined, throw error
-  if (!permissionsCheck) {
-    throwError({ id: "app.no_permissions_defined", data });
-  }
-  // 2. if no document is passed, throw error
-  if (!document) {
-    throwError({ id: "app.document_not_found", data });
-  }
-
-  if (typeof permissionsCheck === "function") {
-    allowOperation = permissionsCheck(options);
-  } else if (Array.isArray(permissionsCheck)) {
-    allowOperation = isMemberOf(user, permissionsCheck, document);
-  }
-
-  // 3. if permission check is defined but fails, disallow operation
-  if (!allowOperation) {
-    throwError({ id: "app.operation_not_allowed", data });
-  }
-};
 
 interface MutationOptions {
   create?: boolean;
@@ -89,7 +37,7 @@ interface BuildDefaultMutationResolversInput {
   options?: MutationOptions;
 }
 
-interface GetMutationDocumentInput {
+interface GetDocumentSelectorInput {
   // TODO: put in common with the single resolver variables type, that have the same fields
   variables: {
     _id: string;
@@ -99,12 +47,30 @@ interface GetMutationDocumentInput {
   context: any;
 }
 
-// get a single document based on the mutation params
+const getDocumentSelector = async ({
+  variables,
+  model,
+  context,
+}: GetDocumentSelectorInput): Promise<{ selector: Object; }> => {
+  const { _id, input } = variables;
+  let selector ;
+  // _id bypass input so we return an empty object
+  if (_id) {
+    return { selector }
+  }
+  
+  const connector = getModelConnector(context, model);
+  const filterParameters = await connector._filter(input, context);
+  selector = filterParameters.selector;
+  return { selector };
+};
+
+/* // get a single document based on the mutation params
 const getMutationDocument = async ({
   variables,
   model,
   context,
-}: GetMutationDocumentInput): Promise<{
+}: GetDocumentSelectorInput): Promise<{
   selector: Object;
   document?: VulcanDocument;
 }> => {
@@ -122,7 +88,8 @@ const getMutationDocument = async ({
     document = await connector.findOne(selector);
   }
   return { selector, document };
-};
+}; */
+
 /*
 
 Default Mutations
@@ -145,15 +112,6 @@ export function buildDefaultMutationResolvers({
       name: getCreateMutationName(typeName),
       async mutation(root, { input: { data } }, context: ContextWithUser) {
         const model = getModel(context, typeName);
-        const { currentUser } = context;
-
-        performMutationCheck({
-          user: currentUser,
-          document: data,
-          model,
-          context,
-          operationName: "create",
-        });
 
         return await createMutator({
           model,
@@ -172,25 +130,16 @@ export function buildDefaultMutationResolvers({
       name: getUpdateMutationName(typeName),
       async mutation(root, { input }, context: ContextWithUser) {
         const model = getModel(context, typeName);
-        const { currentUser } = context;
         const data = input.data;
         const _id = input.id || (data && typeof data === "object" && data._id); // use provided id or documentId if available
 
-        const { document, selector } = await getMutationDocument({
+        const { selector } = await getDocumentSelector({
           variables: {
             input,
             _id,
           },
           model,
           context,
-        });
-
-        performMutationCheck({
-          user: currentUser,
-          document,
-          model,
-          context,
-          operationName: "update",
         });
 
         // call editMutator boilerplate function
@@ -198,55 +147,56 @@ export function buildDefaultMutationResolvers({
           model,
           selector,
           data,
+          dataId: _id,
           currentUser: context.currentUser,
           validate: true,
           context,
-          // document,
         });
       },
     };
   }
 
-  if (mutationOptions.delete) {
-    mutations.delete = {
-      description: `Mutation for deleting a ${typeName} document`,
-      name: getDeleteMutationName(typeName),
-      async mutation(root, { input, _id }, context) {
-        const model = getModel(context, typeName);
-        const { currentUser } = context;
+  // if (mutationOptions.delete) {
+  //   mutations.delete = {
+  //     description: `Mutation for deleting a ${typeName} document`,
+  //     name: getDeleteMutationName(typeName),
+  //     async mutation(root, { input, _id }, context) {
+  //       const model = getModel(context, typeName);
+  //       const { currentUser } = context;
 
-        const { document /*selector*/ } = await getMutationDocument({
-          variables: {
-            input,
-            _id,
-          },
-          model,
-          context,
-        });
+  //       let document ;
+  //       const { document /*selector*/ } = await getMutationDocument({
+  //         variables: {
+  //           input,
+  //           _id,
+  //         },
+  //         model,
+  //         context,
+  //       });
 
-        performMutationCheck({
-          user: currentUser,
-          document,
-          model,
-          context,
-          operationName: "delete",
-        });
-        if (!document?._id)
-          throw new Error(
-            "Perform mutation check did not catch an empty document._id during a delete mutation"
-          ); // should not happen with the check, defnesive code
+  //       performMutationCheck({
+  //         user: currentUser,
+  //         document,
+  //         model,
+  //         context,
+  //         operationName: "delete",
+  //       });
+  //       if (!document?._id)
+  //         throw new Error(
+  //           "Perform mutation check did not catch an empty document._id during a delete mutation"
+  //         ); // should not happen with the check, defnesive code
 
-        return await deleteMutator({
-          model,
-          selector: { _id: document._id },
-          currentUser: context.currentUser,
-          validate: true,
-          context,
-          // document,
-        });
-      },
-    };
-  }
+  //       return await deleteMutator({
+  //         model,
+  //         selector: { _id: document._id },
+  //         currentUser: context.currentUser,
+  //         validate: true,
+  //         context,
+  //         // document,
+  //       });
+  //     },
+  //   };
+  // }
 
   return mutations;
 }
